More fixes
This is now for sure going to fix hoppers entirely, except speeds which we should see in other patch.
Apparently caching the inventory doesn't work for composters as they're constantly updating their block state.
The code now makes it so that we only cache if the block is NOT a composter.

This is also fixing a bug with offline mode servers and proxies for vanilla commands. Read the new patch's patch notes to see what I'm talking about.