Remove Origami's hopper optimizations
See issue #176
This fixes the "it gets more serious". However, the problem is still being present.
it looks like it is purely random which hopper is going to suck in the item.

whatever weirdness is still present with hoppers and sorting systems, if its not mojang
then its paper's optimizations, which we're not removing.