Some small patches
Fixes an issue with limits being ignored with per player spawning
Fixes an issue where you would be respawned south no matter the rotation of your spawn point.

Additionally, removed Producer#fillList(Producer, List) from the collisions patch and moved it
where it is first used. That's a misc maintainability change, but we have to be consistent!