Port 2 patches
First patch is "Do not update distance map when animal and mob spawning is disabled in the particular world"
The commit name speaks for the patch itself. It can decrase tps losses for worlds where animal or monsters do not
spawn.

Second patch is from Purpur, "Spread out and optimise player list ticks".
It changes the way an update packet is sent. The update packet is the packet which updates the green lines
on the tablist.